fixed copilot nitpicks
